avoid partially constructed DifferentiabilityInfo
It's a code smell to make a frozen object and then make a new one with
only a single field modified. It's better to build it all once.

It's fairly easy to accomplish that with the DifferentiablityInfo
objects. We can unconditionally create all the op names and decide to
use it or not when we have all the info needed to make the decision.

ghstack-source-id: 3f4149343574e2c431e1e86255367191abaa81d4
Pull Request resolved: https://github.com/pytorch/pytorch/pull/64238