turn on -Werror=unused-variable in our Bazel CPU build
Summary:
We also fix any existing issues. Note that we only do this for the CPU
build because nvcc is considered a C++ toolchain but it does not have
the same flag support. Adding flags to the GPU build will cause nvcc
errors.

Test Plan: Built locally, rely on CI to confirm.

Reviewers: malfet

Subscribers:

Tasks:

Tags:

ghstack-source-id: 59e9b28c9fa6f0628dd83d81ab8aa707d7646f3f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/79156