[c10d] Change APIs in ProcessGroup to accept const std::vector<at::Tensor>&
Summary:
Change c10d APIs in ProcessGroup to accept const std::vector<at::Tensor>&
instead of std::vector<at::Tensor>&. The intention of the APIs is not to
update the vector. Instead it is to update the data of the held tensors.
This change makes this intention more clear.

Test Plan:
Covered by existing tests.

Fixed #80241.

ghstack-source-id: 23b119b2d4130e3c863b100f8fcfc631e42b41f9
Pull Request resolved: https://github.com/pytorch/pytorch/pull/80996