Stash torch dispatch mode in PythonTLSSnapshot
Previously PythonTLSSnapshot would record only the dispatch key set, but
not the dispatch mode callback. So when using TorchDispatchMode with
torch.override.enable_reentrant_dispatch, the following would happen:

1. Enable TorchDispatchMode, which sets the dispatch mode callback and
enables the Python/PythonTLSSnapshot keys in TLS dispatch key set.
2. Call some torch ops
3. Dispatch to PythonTLSSnapshot, which locally records the dispatch key set
4. Dispatch to Python, which unsets the callback and removes Python from
the TLS dispatch key set
5. Call into the torch_dispatch implementation in python
6. Enable torch.override.enable_reentrant_dispatch, which resets the TLS
dispatch key set that was saved in step 3 (including turning on the Python bit)
7. Dispatch to PythonTLSSnapshot
8. Dispatch to Python (because python bit was set in 6). But, because
the callback was unset in step 4, this errors out.

This will also stash the torch_dispatch callback during
PythonTLSSnapshot, so that we can automatically reset the torch_dispatch
callback when we enable reentrant dispatch.

ghstack-source-id: 6c62cbd53d1fb16b8596c0096b99a09c88a1d847
Pull Request resolved: https://github.com/pytorch/pytorch/pull/81599