[JIT] fix continue + with
Fixes #77354

There was a bug in handling loop continuation + with, where
inlineConsecutiveIfs would try to delete if nodes after the exit
transformations. Previously there would values defined within if blocks
that were *not* returned, but which were still used later in the graph.
When inlineConsescutiveIfs tried to delete the if node it would replace
the output uses and then delete the if nodes; but since some values were
defined in the if block and not returned as an output, the deletion
would fail.

This fixes the issue by registering values as outputs of prim::With if
the values are used later in the graph.

ghstack-source-id: f1d96f6a0ff516c7c31fb6e34b83ea023bff75e5
Pull Request resolved: https://github.com/pytorch/pytorch/pull/79749