Allow MIN_INT in SymInt.
This is a very slight pessimization, as check_ranges() can no longer be
done with a single arithmetic op; instead it's a mask and equality test.

I also took the opportunity to carefully audit SymInt and add more debug
asserts where it might be profitable.  I also simplified mobile macros
where the optimizer should be able to DCE.

Signed-off-by: Edward Z. Yang <ezyangfb.com>

ghstack-source-id: 4ab89220afe4413a6ef27f886dc26197e402d578
Pull Request resolved: https://github.com/pytorch/pytorch/pull/86230