Optimized version of SymIntArrayRef
Previously, to do a conversion of SymIntArrayRef to IntArrayRef, you
always had to a loop over the elements checking if any of them were
symbolic.  Now, at the cost of an extra boolean field on SymIntArrayRef,
we have a bool field you can check directly to see if any element
of SymIntArrayRef is symbolic.  This optimization pays off for eager mode,
as our dispatcher typically deals in SymIntArrayRef and then converts
to IntArrayRef when we call into the C++ kernels.

The cost is in code: I have to duplicate the ArrayRef implementation for
SymInt and do a bunch of modifications.  There is a lot of code and
unfortunately you have to review it all. The main changes:

- Tracking and propagating of AnySymbolic
- A lot of constexpr removal

Signed-off-by: Edward Z. Yang <ezyangfb.com>

ghstack-source-id: bed5507317d43a2a7e6bfa8962e427cb4faa71e2
Pull Request resolved: https://github.com/pytorch/pytorch/pull/86233