Fix ref / decomposition index_add
The decomposition of `index_add` was using `slice(None)`, when it should
use just `None`.

The reference for index_add was also wrong, as `x[idx] += t` does not
use atomic add, so it does not work when several `idx`s point to the
same location.

This PR adds extra reference inputs to help test for this.

Fixes https://github.com/pytorch/torchdynamo/issues/1356

ghstack-source-id: ef4301abbaf4c47ac3a88b5053938f578a874330
Pull Request resolved: https://github.com/pytorch/pytorch/pull/86266