Simplify a few diagonal-related functions
`diag` was unnecessarily implemented as a kernel rather than as a composite
function, which made it unnecessarily difficult (explicit backward + all it entails).

We also change a few uses of `diag` on 2D tensors for `diagonal()`. The
latter returns a view rather than creating a new tensor.

We also upgrade its meta implementation to a fully-fledged
decomposition

ghstack-source-id: a07287c673daae9d2bbfe4aece6d7cb4f043015f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/87180